requirement.txt
# Core Data Science & Numerical Computation
pandas
numpy
scikit-learn

# Machine Learning & Modeling
lightgbm
imbalanced-learn # SMOTEENN 등 불균형 데이터 처리를 위해 필요

# Hyperparameter Optimization
optuna

# Parallel & High-Performance Computing
dask[distributed] # Dask 클라이언트 및 분산 처리를 위해 필요
psutil          # 시스템 CPU, 메모리 등 자원 모니터링

# Model Interpretation & Analysis
shap # SHAP (SHapley Additive exPlanations) 분석

# Data Source
pykrx # 한국 주식 데이터 수집

# Visualization
matplotlib
seaborn

# Utilities & I/O
tqdm # 진행률 표시 바
tabulate # 터미널에 표(table) 형식 출력을 위해 필요
ruamel.yaml # YAML 설정 파일의 주석 및 구조를 보존하며 읽고 쓰기 위해 필요
PyYAML # 일부 모듈에서 표준 YAML 라이브러리 사용
python-json-logger # JSON 형식의 구조화된 로깅
openpyxl # pandas에서 .xlsx 파일(거래내역 분석)을 저장하기 위해 필요
joblib # ML 모델 및 기타 객체 직렬화 (저장/로드)